CrudlDocument component and tests
