+++ src/vuex/modules/common/crudlDocument/mutation-types.js
@@ -1,4 +1,3 @@
-export const DELETE_DOCUMENT = 'DELETE_DOCUMENT'
 export const DELETE_DOCUMENTS = 'DELETE_DOCUMENTS'
 export const RECEIVE_DOCUMENTS = 'RECEIVE_DOCUMENTS'
 export const TOGGLE_SELECT_DOCUMENT = 'TOGGLE_SELECT_DOCUMENT'
