+++ test/unit/specs/crudlDocument/getters.spec.js
@@ -1,4 +1,4 @@
-import {basicFilter, rawFilter, sorting} from '../../../../src/vuex/modules/list/getters'
+import {basicFilter, rawFilter, sorting} from '../../../../src/vuex/modules/common/crudlDocument/getters'
 
 describe('List getters tests', () => {
   describe('basicFilter tests', () => {
